Make comparison const to fix clang c++ warning
In the standard library implementation that comes with the clang toolchain on macOS, there is a `static_assert` that checks for constness in the given multiset's comparator `operator()` implementation. This gets triggered before, when used indirectly, e.g. using the `sort` member function of `tree<>` with a lambda comparison argument. The warning is fixed by adding the `const` keyword.